# packages/backend/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Montessori ERP & LMS API
  description: |
    Multi-tenant, role-based ERP & Learning Management System for Montessori schools.
    
    ## Authentication
    All routes (except `/api/v1/auth/*`) require a JWT Bearer token in the `Authorization` header.
    
    ## Roles
    - **admin**: Full access to all modules
    - **teacher**: Students, attendance, observations, curriculum, messages
    - **parent**: Read-only access to own children, fees, messages
    - **student**: Read-only curriculum and announcements
  version: 1.0.0
  contact:
    name: Montessori ERP Support
servers:
  - url: http://localhost:4001
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
...
paths:
...
  /api/v1/finance/fee-structures:
    get:
      tags: [Finance]
      summary: List fee structures
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Fee structures list
    post:
      tags: [Finance]
      summary: Create fee structure
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [grade, term, amount]
              properties:
                grade:
                  type: string
                term:
                  type: string
                amount:
                  type: number
      responses:
        '201':
          description: Fee structure created
  /api/v1/finance/invoices:
    get:
      tags: [Finance]
      summary: List invoices
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, PAID, OVERDUE, CANCELLED]
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Invoices list
    post:
      tags: [Finance]
      summary: Create invoice
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [studentId, amount, dueDate]
              properties:
                studentId:
                  type: string
                feeStructureId:
                  type: string
                amount:
                  type: number
                dueDate:
                  type: string
                  format: date
      responses:
        '201':
          description: Invoice created
  /api/v1/finance/invoices/student/{studentId}:
    get:
      tags: [Finance]
      summary: Get invoices for a student
      security:
        - bearerAuth: []
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Student invoices
  /api/v1/finance/payments:
    post:
      tags: [Finance]
      summary: Record payment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [invoiceId, amount, method]
              properties:
                invoiceId:
                  type: string
                amount:
                  type: number
                method:
                  type: string
                  enum: [CASH, CARD, BANK_TRANSFER, OTHER]
      responses:
        '201':
          description: Payment recorded

  # ==================== HR ====================
...
  /api/v1/gamification/badges:
    get:
      tags: [Gamification]
      summary: List badges
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Badges list
    post:
      tags: [Gamification]
      summary: Create badge
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                icon:
                  type: string
      responses:
        '201':
          description: Badge created
  /api/v1/gamification/badges/{badgeId}/award/{studentId}:
    post:
      tags: [Gamification]
      summary: Award badge to student
      security:
        - bearerAuth: []
      parameters:
        - name: badgeId
          in: path
          required: true
          schema:
            type: string
        - name: studentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Badge awarded
  /api/v1/gamification/points:
    post:
      tags: [Gamification]
      summary: Award points to student
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [studentId, value, reason]
              properties:
                studentId:
                  type: string
                value:
                  type: integer
                reason:
                  type: string
      responses:
        '201':
          description: Points awarded
  /api/v1/gamification/points/student/{studentId}:
    get:
      tags: [Gamification]
      summary: Get student points
      security:
        - bearerAuth: []
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Student points
  /api/v1/gamification/leaderboard:
    get:
      tags: [Gamification]
      summary: Get leaderboard
      security:
        - bearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            default: weekly
      responses:
        '200':
          description: Leaderboard entries
  /api/v1/gamification/leaderboard/calculate:
    post:
      tags: [Gamification]
      summary: Recalculate leaderboard
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                period:
                  type: string
      responses:
        '200':
          description: Leaderboard recalculated

  # ==================== AI ====================
...